Migrate to BigDecimal

+++ src/main/kotlin/kuants/Quantity.kt
@@ -1,17 +1,22 @@
 package kuants
 
+import java.math.BigDecimal
+import kotlin.math.minus
+import kotlin.math.times
+import kotlin.math.plus
+
 /**
  * Created by evacchi on 18/07/14.
  */
 trait Quantity<T : Quantity<T>> {
-    val value: Double
+    val value: BigDecimal
     val valueConstructor: UnitOfMeasure<T>
     val valueUnit: UnitOfMeasure<T>
 
-    fun to(u: UnitOfMeasure<T>): Double = u.converterTo(value)
+    fun to(u: UnitOfMeasure<T>): BigDecimal = u.converterTo(value)
 
     override fun toString() = "${value} ${valueUnit.symbol}"
     fun plus(that: Quantity<T>): Quantity<T> = valueUnit(this.value + that.value)
     fun minus(that: Quantity<T>): Quantity<T> = valueUnit(this.value - that.value)
-    fun times(scalar: Double): Quantity<T> = valueUnit(this.value * scalar)
+    fun times(scalar: Double): Quantity<T> = valueUnit(this.value * BigDecimal(scalar))
 }
