+++ src/main/kotlin/kuants/mass/Mass.kt
@@ -3,18 +3,19 @@ package kuants.mass
 import kuants.Quantity
 import kuants.UnitOfMeasure
 import kuants.MetricSystem
+import java.math.BigDecimal
 
 /**
  * Created by evacchi on 18/07/14.
  */
-class Mass(override val value: Double) : Quantity<Mass> {
+class Mass(override val value: BigDecimal) : Quantity<Mass> {
     override val valueUnit = Grams
     override val valueConstructor = Grams
     fun invoke(value: Double) = valueConstructor(value)
 }
 
 trait MassUnit : UnitOfMeasure<Mass> {
-    override fun invoke(value: Double) = Mass(value)
+    override fun invoke(value: BigDecimal) = Mass(value)
 }
 
 object Kilograms: MassUnit {
